Finished all checkpoints in Chapter 4, entered Case Studies

+++ Liang/Chapter4/CheckPoint4_14.java
@@ -2,6 +2,6 @@ public class CheckPoint4_14
 {
   public static void main(String[] args)
   {
-    System.out.println((char)((int)Math.floor(Math.random() * ('z' - 'a')) + 'a'));
+    System.out.println((char)Math.floor(Math.random() * ('z' - 'a') + 'a'));
   }
 }
